Auth merge tested to work with old and new frontend
To configure the frontend, point cgi script to the backend and
uncomment the login option in index.html; to configure the backend,
point AUTH_SERVER to auth.cgi (sibling to korp.cgi and callable in
localhost, so may just work).

To use in another server, prepare the database with
src/authing/table-auth.sql and then corpora and users with
src/authing/auth; public (PUB) corpora need not be in the database,
restricted (ACA and RES) corpora also need configure with
limited_access : true in frontend.