korpplugins: Extract library code to korppluginlib
Extract the Korp plugin library code (and its documentation) in package
korpplugins to a new package korppluginlib.

In this way, korpplugins need not have an __init__.py and can be a
namespace package, whose modules or subpackages can reside in different
directories, so that plugins need not be under the same directory
structure as the Korp backend.

The code was not yet changed to reflect the move, so it does not work
yet.